Move file and use a wrapped fopen instead of directly
